requirements.txt
ipywidgets==8.1.2
matplotlib==3.8.4
pandas==2.2.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.4
fonttools>=4.60.0
filelock>=3.20.3

fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
sqlalchemy==2.0.21
pydantic==1.10.12
python-dotenv==1.0.0
pytest==7.4.0
respx==0.20.0
dnspython==2.4.2
pika==1.3.1
aiohttp==3.8.5
phonenumbers==8.13.19
requests==2.31.0
rapidfuzz
sentence-transformers
